Made a decomposition of the source files.
