+++ CMakeLists.txt
@@ -8,12 +8,12 @@ find_package(Threads REQUIRED)
 find_package(OpenSSL REQUIRED)
 find_package(Boost COMPONENTS system REQUIRED)
 find_package(CURL)
-include_directories(/usr/local/include ${OPENSSL_INCLUDE_DIR} ${Boost_INCLUDE_DIR} include/)
+include_directories(inc /usr/local/include ${OPENSSL_INCLUDE_DIR} ${Boost_INCLUDE_DIR} include/)
 if (CURL_FOUND)
     include_directories(${CURL_INCLUDE_DIRS})
     add_definitions(-DHAVE_CURL)
 endif()
-add_executable(tgbot main.cpp)
+add_executable(tgbot src/bot_api.cpp main.cpp)
 
 target_link_libraries(tgbot /usr/lib/libTgBot.so ${CMAKE_THREAD_LIBS_INIT} ${OPENSSL_LIBRARIES} ${Boost_LIBRARIES} ${CURL_LIBRARIES})
