+++ inc/bot_api.h
@@ -1,0 +1,9 @@
+/*
+    ShellBot
+    Created by alimovlex.
+    Copyright (c) 2020 alimovlex. All rights reserved.
+*/
+#include <tgbot/tgbot.h>
+using namespace TgBot;
+
+void runBot(Bot &build);
